ganon_client/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ganon-client"
version = "0.1.0"
description = "Python client library for ganon mesh network tunneler"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = ["construct>=2.10", "monocypher-py>=0.0.3"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ganon_client"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
